Added eslint to the project
command used to install eslint:

    npm install eslint --save-dev

eslint will be useful mostly for keeping the code style more uniform,
and to catch anything strange happening that is easily fixable.

For example, eslint can be used to simply format the code, making
sure there are 4 spaces per tab, and that the code only has actual
spaces and not "\t" characters.